+++ python/spglib/__init__.py
@@ -38,6 +38,8 @@ from ._version import __version__, __version_tuple__  # noqa: F401
 
 # fmt: off
 from .spglib import (  # noqa: F401
+    MagneticSpaceGroupType,
+    SpaceGroupType,
     delaunay_reduce,
     find_primitive,
     get_BZ_grid_points_by_rotations,
@@ -48,6 +50,7 @@ from .spglib import (  # noqa: F401
     get_ir_reciprocal_mesh,
     get_layergroup,
     get_magnetic_spacegroup_type,
+    get_magnetic_spacegroup_type_from_symmetry,
     get_magnetic_symmetry,
     get_magnetic_symmetry_dataset,
     get_magnetic_symmetry_from_database,